Build `MachineRegister` on `CapabilityPtr`.
Previously, `CapabilityPtr` wrapped `MachineRegister`. This initially made sense because every `CapabilityPtr` is logically a `MachineRegister` but not every `MachineRegister` is a `CapabilityPtr`. However, we've realized that building `CapabilityPtr` atop `MachineRegister` requires `MachineRegister` to expose all of `CapabilityPtr`'s functionality, which is not correct for `MachineRegister`. Therefore, this commit reverses that relationship: `MachineRegister`'s implementation now wraps `CapabilityPtr`.

There is no conceptual change in what the two types mean. However, `CapabilityPtr` now promises that it can contain an integer for `MachineRegister`'s use -- but other code is not allowed to rely on that.