chips: stm32wle5xx: update int. handler; add lora
This adds the subghzspi peripheral and wrapper for virtual gpios to the
subghz interrupt / busy signals to the stm32wle5xx chip definition.

Because the subghzradio interrupts are level triggered and can only be
cleared by initiating spi commands, we mask the subghz radio interrupts
when determining if the chip has pending interrupts and when finding the
next pending interrupt.